Reduce logic in Organization model
